moving folders to typescript example directory
