+++ examples/typescript/l1-resources/auditing.ts
