+++ examples/typescript/l1-resources/cloud-backup-snapshot-export-bucket.ts
