+++ examples/typescript/l1-resources/cloud-backup-snapshot.ts
