+++ examples/typescript/l1-resources/custom-db-role.ts
