+++ examples/typescript/l1-resources/custom-dns-configutation-cluster-aws.ts
