+++ examples/typescript/l1-resources/datalake.ts
