+++ examples/typescript/l1-resources/encryption-at-rest.ts
