+++ examples/typescript/l1-resources/federation-settings-org-role-mapping.ts
