+++ examples/typescript/l1-resources/global-cluster-config.ts
