+++ examples/typescript/l1-resources/ldap-configuration.ts
