+++ examples/typescript/l1-resources/ldap-verify.ts
