+++ examples/typescript/l1-resources/network-container.ts
