+++ examples/typescript/l1-resources/network-peering.ts
