+++ examples/typescript/l1-resources/online-archive.ts
