+++ examples/typescript/l1-resources/org-invitation.ts
