+++ examples/typescript/l1-resources/private-endpoint-regional-mode.ts
