+++ examples/typescript/l1-resources/private-endpoint.ts
