+++ examples/typescript/l1-resources/project-invitation.ts
