+++ examples/typescript/l1-resources/project.ts
