+++ examples/typescript/l1-resources/search-index.ts
