+++ examples/typescript/l1-resources/serverless-instance.ts
