+++ examples/typescript/l1-resources/teams.ts
