+++ examples/typescript/l1-resources/third-party-integration.ts
