+++ examples/typescript/l1-resources/trigger.ts
