+++ examples/typescript/l2-resources/atlas-encryption-at-rest.ts
