+++ examples/typescript/l2-resources/third-party-integration.ts
