+++ examples/typescript/l3-resources/atlas-basic-private-endpoint.ts
