+++ examples/typescript/l3-resources/atlas-basic.ts
