+++ examples/typescript/l3-resources/encryption-at-rest-express.ts
